[move] functionを分割
