+++ lib/src/commonMain/kotlin/com/wsr/Serializer.kt
@@ -4,9 +4,9 @@ import com.wsr.common.IOType
 import com.wsr.layers.Layer
 import com.wsr.layers.affine.AffineD1
 import com.wsr.layers.bias.BiasD1
-import com.wsr.layers.function.ReluD1
-import com.wsr.layers.function.SigmoidD1
-import com.wsr.layers.function.SoftmaxD1
+import com.wsr.layers.function.relu.ReluD1
+import com.wsr.layers.function.sigmoid.SigmoidD1
+import com.wsr.layers.function.softmax.SoftmaxD1
 import com.wsr.layers.reshape.ReshapeD2ToD1
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.descriptors.SerialDescriptor
