+++ lib/src/commonMain/kotlin/com/wsr/layers/function/sigmoid/SigmoidD1.kt
@@ -1,4 +1,4 @@
-package com.wsr.layers.function
+package com.wsr.layers.function.sigmoid
 
 import com.wsr.NetworkBuilder
 import com.wsr.common.IOType
